refactor: moved interface + changed exports
